Add WebAssembly as Intrinsic, not module.
Fix some warning.

+++ quickjs.c
@@ -10728,7 +10728,7 @@ static int JS_ToInt64SatFree(JSContext *ctx, int64_t *pres, JSValue val)
             } else {
                 if (d < INT64_MIN)
                     *pres = INT64_MIN;
-                else if (d > INT64_MAX)
+                else if ((long)d > INT64_MAX)
                     *pres = INT64_MAX;
                 else
                     *pres = (int64_t)d;
@@ -53865,7 +53865,7 @@ static JSValue js_atomics_wait(JSContext *ctx,
     }
     if (JS_ToFloat64(ctx, &d, argv[3]))
         return JS_EXCEPTION;
-    if (isnan(d) || d > INT64_MAX)
+    if (isnan(d) || (long)d > INT64_MAX)
         timeout = INT64_MAX;
     else if (d < 0)
         timeout = 0;
